requirements.txt
# Core
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0

# Auth
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# Environment
python-decouple==3.8

# Utilities
django-extensions==3.2.3
Pillow==10.4.0

# Dev / Testing
pytest==8.3.2
pytest-django==4.8.0
factory-boy==3.3.1
